Run in a `nix` shell
We want to have reproducibility so we want to run inside of `nix`.

Previously, `bazel` was downloading itself. Not entirely sure how that
was happening. Maybe `bazel` in a GitHub Action is actually `bazelisk`?
Either way, we don't want to use that directly.

We pull in the `workflow/nix-shell-action` and use it to run inside of a
`nix` shell.