Initial bootstrapping
We setup just the very absolute minimum files and expand on the README
just a touch. This should be enough to get us rolling.

We use `nix` to boostrap everything, then `bazel` as the workhorse of
the build. We'll probably want to add some simpler ways to build
everything when it starts to get complex. But we have time before that
happens.

We also setup a GitHub Action to run a quick job that should validate
that everything is working.

Branch: joneshf/initial-bootstrapping
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/1