Update nix to latest version of 22.11 channel
We're using the release channel of 22.11 because it should be a stable
channel. There have been bug fixes and other updates since the previous
version. We update so we have the latest.

Two of the things that got updated (that are relevant to us) are:
1. We moved `bazel` from `6.0.0-pre.20220720.3` to `6.0.0`. The
    pre-released version worked, but it has some differences from what
    was released in `6.0.0`.
2. We moved `go` from `1.19.3` to `1.19.5`. We're not yet using `go`
    directly, so it's not a big deal. But it's good to note.

Branch: joneshf/update-nix-to-latest-version-of-22-11-channel
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/2