Remove `bazel` for `make`
Turns out, trying to get `bazel` working with `go` protobufs is
non-trivial on a fairly standard macOS device. It seems that it wants a
bunch of headers to be available globally for compiling protobufs, and I
don't have the time to figure out what all that is. I sunk more than a
day into diagnosing it, and it's not easy to figure out.

If at some point this becomes easier, we can try again.