Setup a minimal provider
We start by setting up a very minimal provider. So minimal that it
actually doesn't do anything (and likely wouldn't work if we wrote a
Terraform file attempting to use it). We'll flesh this out more over
time. But this ought to get us rolling for now.

We also setup documentation so we at least know that we'll have that
working for us. It's not a whole lot (again, because it's so minimal),
but it gets us started down the right path.