Ignore timestamp differences
When we run `make docs`, it'll overwrite the timestamp of files. And
apparently, that's a thing that `git diff-index` will pick up as a
difference.

We make `git` ignore this by first updating the index. It'd be nice if
there were a flag we could pass to `git diff-index` that would allow it
to ignore the timestamps. It doesn't seem like there is one though.