Fix docs "up to date" test
Before, when we had `git update-index`, it would error out for any
changes at all. Which meant that even if you were just running `make
test` for a simple thing not related to docs, it would still error out.
We don't want this behavior.

Despite trying a bunch of different things, we couldn't figure out how
to get it to not error on this line. This seems to stem from `errexit`
being an extremely convoluted thing:
https://mywiki.wooledge.org/BashFAQ/105.

In any case, we switch around what we do so that instead of dealing with
`git update-index`, we deal with `git status` directly in the
conditional. This only works because of the complexities behind
`errexit`. So really, we need to just drop `errexit`…

Branch: joneshf/fix-docs-up-to-date-test
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/5