Start fleshing out the OpenWRT provider
This is kind of a big chunk of changes. The long and short is that we
should be able to start writing provider stuff for real now.

The tests leave something to be desired. And it looks like there's not
really a great way to unit test this stuff. We tossed together a few
simple tests, but nothing really great. In particular, it seems
incredibly hard to validate the schema. Hopefully, we can figure out how
to do that at some point.

Branch: joneshf/start-fleshing-out-the-openwrt-provider
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/6