Start working on an RPC client for LuCI
Apparently, LuCI has an RPC client that can be used a little easier than
`go-uci`: https://github.com/openwrt/luci/wiki/JsonRpcHowTo#basics. It
requires installing a couple of packages, so it might not work for
everyone. But, it'll work for us.

We toss together a simple package that makes setting up and getting data
from a section relatively straight-forward. We'll add more behavior as
we need it. But, this is a decent start.