Hold onto the JSON-RPC client
If we separate creation from invocation, we can end up with a simpler
situation overall. Instead of needing to pass around an HTTP client to
each method, along with an address, we can just keep the JSON-RPC client
for the specific endpoint and invoke it later with the method we want.