Extract JSON-RPC implementation
We pull this stuff out so the rest is a little easier to understand.
It's very unfortunate that we even need to write this code, but the
`net/jsonrpc` package doesn't work with more than one value in
`"params"`, so here we are.

Branch: joneshf/extract-json-rpc-implementation
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/8