Test against an actual device
We were kind of going by our best guess before with the behavior. Since
things can sometimes be out of sync, we want to run some tests against
the real thing. We setup a quick `docker` situation using the 22.03.3
version, and test against that. Thankfully, there are images pre-built!

These tests fleshed out some issues already. In particular, the format
of the response is not at all what we expected. Sometimes it comes back
with `"result"` being an object, sometimes it's an array, likely it can
be other things when we invoke other commands. We adjust the
implementation to handle these cases.

The tests we add for acceptance are given a build tag of `acceptance` so
that we can reference them easily. We also add a couple of `make`
recipes to make the different testings easier. Though it's a shame that
there's no easy way to only run the acceptance tests.

Finally, to make sure everything can actualy work no matter what people
have installed, we add `colima` to the nix environment. This program
lets us run a `docker` daemon on the fly, instead of requiring everyone
setup `docker` on their system.