Make acceptance test build constraint explicit
Bytes are cheap these days, we don't have to be terse or cryptic with
the build constraints for acceptance tests. Instead we can give them a
more descriptive name.

The syntax for build constraints seems hard to discover, but it looks
like the convention for non-alphanumeric characters is to use `.`.