Switch over to using LuCI's JSON-RPC interface
The UCI interface requires being able to run commands on files that you
have access to. Currently, it looks like there's no easy way to run ths
on remote files (like a router that isn't on the same system as
`terraform`). And since it's highly unlikely that someone will have
`terraform` installed on their router, this becomes a hard-to-start
situation.

There seem to be a couple of efforts to make it possible to deal with
remote files:
- https://github.com/digineo/go-uci/pull/14
- https://github.com/digineo/go-uci/pull/19

But, both of these appear to be a bit stuck for different reasons. Maybe
they'll land at some point (which would be great).

In the meantime, we can unstick ourselves by using LuCI's JSON-RPC
interface instead. This is arguably simpler anyway because it means the
provider just has to deal with an HTTP(s) interface, like most other
providers.

We swap over to that, add the attributes we need, massage some
implementation details and tests, and we should be on our way. We ended
up converting one of the tests to an acceptance test, so it should give
us some good confidence that we've got things setup properly.