Add a health check for the acceptance test server
We were not waiting for this server to be ready, only starting it and
assuming it was up and running. This lead to tests failing when we made
the dependencies more accurate. So, we add a health check and wait on
it.

This is a good failure, it's made our code base more robust and
descriptive.