Clean up acceptance test recipes
We only need to start the acceptance test server when we want to run
acceptance tests. We make the dependencies reflect that.

We actually flesh out an interesting issue with this change. Before, the
acceptance test server would not be needed until after it was already
ready. This was due to other things running between the time the
acceptance test server was started and the acceptance tests actually
running. Now that we only start the acceptance test when we need it, we
can get into a situation where the acceptance tests are attempting to
run before the server is actually ready.

To address this situation, we add a `HEALTHCHECK` to the `docker`
container, and wait for it to be healthy by passing the `--wait` flag to
`docker compose`. This makes it so that initial start might take a bit
more time than before, but we won't fail the tests when we should've
waited for the acceptance test server to be ready in the first place.

To be clear, this health check business is a good thing. It might make
bringing up the acceptance test server a little slower initially, but
that cost should be cheap in CI, and almost free in development. More
importantly, it means we should have fewer transient errors from not
waiting for the server to be ready.

Branch: joneshf/clean-up-acceptance-test-recipes
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/13