Add a data source for system.system
We want to be able to read the data out of the system for whatever
reason. For instance, we may want to use the `hostname` to interpolate
elsewhere, or use as a value in an entirely different resource.

We toss together a quick data source for the system information so that
we can do that. Eventually, we'll want to have a resource for
system.system, but this is a good start for at least getting some of
OpenWrt communicating with other Terraform infrastructure.