Setup releases
There's a good chance we're going to have to iterate on this after we
merge it. For now, we're mostly following the tutorial:
https://developer.hashicorp.com/terraform/tutorials/providers/provider-release-publish,
but we're calling an audible on the releaser action. Since we're fully
vested in `nix`, we try to release through `nix`. But, there's no
telling whether this will work. I guess we'll see soon…