Fix part of release process
These are things that showed up when running locally. Hopefully there's
not much more when running as a GitHub action.

The first issue is that the version of `goreleaser` we're using (1.13.0)
doesn't support a `--clean` flag on `goreleaser release --clean`.
`--clean` was released in 1.15.0. The equivalent flag is `--rm-dist`, so
we use that instead.

Next the arguments we pass to `gpg` for signing were off. We re-ordered
them, but `gpg` doesn't actually work like that. Some of the flags are
actually commands, and if they're not in the right order, everything
breaks.

Finally, we fix a little copy/paste issue we had for one of the steps in
the release job.

Branch: joneshf/fix-part-of-release-process
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/17