Don't worry about draft releases
It's an extra step we don't really need to think about. A better
situation would be to have a release always ready to go and just publish
when it made sense. We'll get there eventually, but not at the moment.

Branch: joneshf/don-t-worry-about-draft-releases
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/18