Add the version to provider
This might be the reason the provider isn't really working. Or it could
be that the version is 0.0.0. Either way, we're adding it now. This
should get altered by `goreleaser` when it runs since it passes in
linker flags to change the `main.version` value.

Branch: joneshf/add-the-version-to-provider
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/20