Remove Data Source naming where possible
There's nothing about these names that's specific to the Data Source. It
was just named this way because it seemed like defaulting to more
verbose names would be good.

Now that we're moving on to wanting a Resource, it's more consistent to
have the same stuff used both for the Data Source and the Resource.

We remove the Data Source naming where we can to make it less confusing
in the long run.