Return diagnostics from most helpers
Instead of mutating a `datasource`-specific type in each of these
helpers, we flip things around and just return the diagnostics directly.
This puts the onus on the caller to do the massaging of the diagnostic,
but it makes it so the helpers are no longer dependent on a specific
implementation.

This should allow us to extract these helpers to a place that can be
used by more than just this one Data Source.