Extract `system.system` attributes
These ought to be the same for both the Data Source that currently
exists and the Resource we're planning to write. In order to only write
these attributes once, we define our own simple types for what
information the attributes need, and provide an interface that each one
can adhere to. We then iterate them convert to the appropriate schema
attribute.

This is a lot of work that it feels like the Terraform framework should
be able to figure out. It's not a terrible amount, but it's a bit
ridiculous that we have to do it anyway.

Branch: joneshf/extract-system-system-attributes
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/32