Make JSON-RPC params `json.RawMessage`
As it turns out, each one of these methods has different parameters. So,
we can't just use a string for all of them. Instead of completely losing
any type safety with `any`, we say that they all have to be
`json.RawMessage`. It's still possible to get wrong, but hopefully not
too easy.