Add support for creating a section in `lucirpc`
We want to be able to actually create things on an OpenWrt device. We
add this `CreateSection` method so that we can do exactly that.

Since this is where things start to get kind of weird for the LuCI
JSON-RPC API, we write some tests to catalog the behavior. These tests
show a bit of the inconsistency in the API. Thankfully, we've got a
package that tries to smooth some of that out for us.