Support resources from general attributes
We can slightly extend our generalized attributes to support resources
by providing a method that does the conversion. This should make it
possible to have exactly one definition for the attributes instead of
letting them diverge.

This also showcases why we named the `DataSourceExistence` property the
way we did. The two aren't always in line between Data Sources and
Resources, so we allow them to diverge in each attribute.