Move docker stuff to the top-level
It was always a little odd that we were using the docker stuff from
within `lucirpc`. There's no good place for this that works for both
`lucirpc` and `openwrt`, so we're just going to move it to the
top-level. If we find a better place for this, we can move it there.

Branch: joneshf/move-docker-stuff-to-the-top-level
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/35