Run a separate OpenWrt server in each test
Up until now, things have been fine to share one single OpenWrt server
across all of the tests. That's because everything has been read-only.
So, it really didn't matter if each test was hitting the same server or
a different server.

Now that we want to start mutating these OpenWrt servers, sharing one
server across multiple tests would mean that we'd run into race
conditions once we started mutating.

To alleviate that, we run a separte OpenWrt server in each test. This
might seem like overkill, but this should be more reliable in the long
run. We might end up running out of resources by trying to spin up a
docker container for each test, but the actual OpenWrt server is pretty
light weight (it's supposed to run on consumer-grade routers). So, it
will _probably_ be a while before we run into any resource limits. As a
data point, each of these tests seems to only consume 4-6MB of memory
and relatively little CPU.

We setup an internal package that takes care of most of the boilerplate
for us, so we can focus on writing the actual tests. This lets us use it
in multiple different packages fairly easily.

Finally, since we no longer rely on the container setup in
`docker-compose`, we remove that as a dependency in the
`go-test-acceptance-test` recipe.

Branch: joneshf/run-a-separate-openwrt-server-in-each-test
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/36