Make "requiredness" more correct (as "existence")
The Terraform framework does the incorrect thing here and attempts to
represent the four possibilities for requiredness as three `bool`s.
This isn't right because four out of the eight possible combinations are
invalid.  E.g. It errors somewhere if all three `bool`s are `false`, or
if required and something else are `true`.

Instead of inheriting that incorrect behavior, we make our lives simpler
by creating an "enum" of the four possible values. This "enum" is then
used to convert into that invalid representation that Terraform wants.
This way, we have a bit more confidence that we're doing the right
thing.

In trying to figure out what the name of this "enum" is, it seems like
we're more defining how the attribute exists rather than whether or not
it's required. We choose to go with a term that includes "existence."

The word "enum" is used because what we have is an approximation of a
real enum. But, this is an idiomatic pattern in go. Maybe one day we'll
get real support for enums and can have certainty that things won't go
wrong.

Branch: joneshf/make-requiredness-more-correct-as-existence
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/39