Support not finding a section
We don't always know that we'll find a section. Before, we were assuming
that any call to `lucirpc.GetSection` should always succeed. But if we
get a config/section that doesn't exist, we expect not to find anything.

We change to do a little more explicit checking.

Branch: joneshf/support-not-finding-a-section
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/40