Take the section name as an argument
When we get to adding support for Resources, we can no longer assume
we're dealing with a single system.

It's unclear if OpenWrt actually allows this in a running device. But
the LuCI JSON-RPC API definitely allows you to create a new
`system.system` section. So we're going to support it for now.

If we find that this is actually incorrect, we'll remove this behavior.
Just because the LuCI JSON-RPC API allows something doesn't mean this
Terraform provider has to encourage creating a broken OpenWrt device. We
can attempt to smooth out some of the rough edges.

Branch: joneshf/take-the-section-name-as-an-argument
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/41