Add support for creating a section in `lucirpc`
We want to be able to actually create things on an OpenWrt device. We
add this `CreateSection` method so that we can do exactly that.

Since this is where things start to get kind of weird for the LuCI
JSON-RPC API, we write some tests to catalog the behavior. These tests
show a bit of the inconsistency in the API. Thankfully, we've got a
package that tries to smooth some of that out for us.

This also shows the first time that we need to be concerned about
committing changes. When reading, we don't have to worry about that too
much, but the model for UCI is to have a distinction between changes
that are still in progress and changes that have been committed.

We decide not to worry about that (since every one of our consumers
would have to know about and follow this UCI pattern), and make it so
it's one action (at least in theory). If we find that it would be better
to follow this UCI pattern, we can always change the API to make it work
like that.

Branch: joneshf/add-support-for-creating-a-section-in-lucirpc
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/42