Add support for updating a section in `lucirpc`
We want to be able to update sections. This allows us to start managing
resources for real with Terraform. We'll probably need to flesh this out
more once we have a real resource going, but this should be good for
now.

We add a handful of tests to try and catch some of the behavior of this
API. We add both unit tests and acceptance tests.

Branch: joneshf/add-support-for-updating-a-section-in-lucirpc
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/43