Add support for deleting a section in `lucirpc`
With this, we should have most of the behavior for managing a UCI
section as a Terraform resource. This is similar to the other CRUD
operations, so there's not much explanation here.

We add some unit and acceptance tests to validate how the implementation
works.

Branch: joneshf/add-support-for-deleting-a-section-in-lucirpc
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/44