Add support for managing `system.system` Resources
This is our first resource! We start with the `system.system` Resource,
since we already have a Data Source for it. It's not actually likely
that someone will create one of these (as there's only one
`system.system` section in a running OpenWrt device). However, it's very
likely that someone will import one of these and then make changes
afterward.

For instance, wanting to change the timezone, logging, or hostname would
all require importing the current `system.system` section, and then
changing it.

Most of this is boilerplate for setting up the resource. Hopefully as we
get more experience adding resources, we can mitigate how much of this
book keeping we actually need to do. Until this, this approach seems to
work.

Branch: joneshf/add-support-for-managing-system-system-resources
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/46