Move handling of `bool` into glue code
This might be a misstep, but it feels like not being able to create,
delete, or update the section should be an error outright and stuffed
into the `Diagnostics`. We switch to that in hopes that this is easier
in the long run.

Branch: joneshf/move-handling-of-bool-into-glue-code
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/49