Fetch the tags when making a release
We also need more of the history in order to fetch the tags. We default
to fetching all of the history because we don't have that many commits
(and won't for a long time). If it starts to become an issue in a few
years, we can do something differently.