Move helpers into other file
It's a little awkward to have these helpers in the `system_resource.go`
file, since they're only used in the `system.go` file. There's some sort
of implicit hierarchy here, and it was inverted. We move the helpers
around, so it gets back to normal.

This probably should've been done first.