Move create/read/update to each attribute
We can move the logic for doing each of these things into the attributes
instead of having to do them in the Data Source and Resource methods.
This helps clean up those methods, while also co-locating the behavior
of each attribute with the rest of its logic.

Naïvely, it seems like this is how it should be anyway. Each attribute
should know how to perform the operations itself. But, maybe when we get
further into it, we'll find that this doesn't really work out the way
it's supposed to. That likely means there's probably a union type that
can't be expressed, but that's a story for another day.

Branch: joneshf/move-create/read/update-to-each-attribute
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/51