Make helpers for reading/upserting attributes
All of the logic of these functions is effectively the same. We just
have to account for the different types. To that end, we can pull out
the commonalities as helpers and parameterize them with the things that
are different.

We have to write a little boilerplate to make the getters/setters
first-class functions we can pass around, instead of syntax built into
the language. But that's a cheap thing to maintain.

The only one that doesn't really fit is the `Id`. It uses the metadata
instead of the actual data. If we start to see a pattern with metadata,
we can provide a helper for that. Until then, it's probably easier to
just write this one explicitly.