Extract attribute logic to glue package
This is all stuff that can be useful to other Data Sources and
Resources. We pull these out to the glue package, since that seems to be
a good spot for it. It's still a little entangled with the way LuCI's
JSON-RPC API works.

Branch: joneshf/extract-attribute-logic-to-glue-package
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/53