Move terraform types to the "glue" package
These really ought to be in a proper Terraform package. But we stick
them here, since there's no better place for them.

It's entirely possible that these exist somewhere else in the world. In
which case, we could pull them in from there instead of defining them
here ourselves.

Branch: joneshf/move-terraform-types-to-the-glue-package
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/57