Add support for `network.globals`
We add support for both the Data Source and Resource. This stuff is
pretty obscure in LuCI, but still important to have. Packet steering in
particular can really change the performance of the router.

This is a fairly cut-and-dry addition. It's our first foray into the
network config, so it might flesh out some issues we weren't aware of.

Branch: joneshf/add-support-for-network-globals
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/60