Don't use `resource.ParallelTest`
While `resource.ParallelTest` would normally be equivalent to using
`testing.Parallel` followed by `resource.Test`, it's not quite the same
for these test cases.

We do some setup work to have the OpenWrt server running in the test.
Since that setup work hasn't been marked as parallel (by a
`testing.Parallel` call) it runs serially. So we end up with a chunk of
the test cases running serially, and a chunk of them running in
parallel. While that's an interesting observation, that's not what we
want.

Instead, we explicitly mark the start of the test case as parallel with
`testing.Parallel`, and then use the `resource.Test` helper. This lets
the entire test case be run in parallel with the others.

The end result of this change is that instead of each test waiting the
5-10 seconds required for the OpenWrt server to be healthy before
allowing other tests to run, we spin up all of the OpenWrt servers at
the same time (or close enough to it to not matter). The time in the
test suite grows at a constant rate instead of at a linear rate with the
amount of tests.