Support sets of strings as attributes
There are quite a few sets of things in OpenWrt. We need to start
supporting them. One of the soon-to-be-needed attributes is the ports on
a bridge device. Those ports are a set of strings (even if UCI
represents them as a list of strings). So, we add support for attributes
that are sets of strings.

This work will almost surely need to be adjusted as we get more
experience with the attributes. So this work will change over time.

This work also doesn't preclude lists of strings. We just don't have an
immediate need for them at the moment.

Branch: joneshf/support-sets-of-strings-as-attributes
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/65