Support validating other attribute values
There's limited support for validating other attributes. In particular,
it seems the only thing that comes out of the box is support for
validating the existence of another attribute. This seems fairly
unuseful, since an attribute just existing doesn't tell whether or not
the value of that attribute is what it is supposed to be.

We add a validator that not only checks for existence, but also checks
that it has the correct value. This validator should allow expressing
something like: "`foo` must be `true` when this attribute is set," or
"`bar` must be `"baz"` when this attribute is set."

This validator doesn't look at the current attribute's value for
validation. That's another constraint that we don't necessarily need at
the moment. We can add that later on, if it becomes something we want,
though.