Also support validating set and string attributes
We missed this when adding the validator. But, we do want to support the
other types of attributes. Since we also have sets and strings, we add
support for those here.

The fact that there's no actual difference here should be telling that
the underlying abstraction isn't quite right. Probably, we need to our
adapter layer to be more distanced from the Terraform packages. There's
little reason we should have to deal with this much copy/paste just
because the underlying framework made some decisions that so far haven't
been beneficial to us.

Branch: joneshf/also-support-validating-set-and-string-attributes
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/67