Add support for `network.device`
We add support for both Data Sources and Resources. We only add support
for bridge devices. Mostly because that's all I personally need at the
moment. We will add support for the other device types in the future.

The documentation on `network.device` is missing on this day
(2023-03-11):
https://openwrt.org/docs/guide-user/base-system/basic-networking#device_sections.
Most of what's added in this change is based on inferring the behavior
in a running version of OpenWrt 22.03.3. There's a good chance that
we're missing some behavior. Hopefully, we're not missing something that
can cause catastrophic failures. In any case, we'll improve on this as
it goes on.

Meta observation: it seems like we might be coalescing on a pattern for
Data Sources and Resources. It seems most of the behavior can be
expressed in the schema. That's really what we want at the end of the
day anyway. Might try and abstract it so we don't have to write the
couple hundred lines of boilerplate each time.

Branch: joneshf/add-support-for-network-device
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/68