Make helper for reading model
We're doing the same thing in each of these, and there's no indication
that we're going to do something differently in the future. We pull this
logic to a centralized location and use it in the three different spots
we've got.

This isn't that huge of a change, since the duplicated behavior was
already pretty well abstracted, but it's a step to get us to seeing only
the differences rather than the similarities.