Grab the id outside of generating the API body
We don't actually need to return the id as part of the generation of the
API body. Each call-site has the model already (it's what the method is
being called on). Instead of doing that, we can grab the id outside of
the method when we need it.

This simplifies the API of the method a bit, but also makes it so we can
drop the hard requirement on having a model with this particular shape
of `Id` with a `ValueString` method while generating the API body. It
was a bit over-specified to begin with.

Branch: joneshf/grab-the-id-outside-of-generating-the-api-body
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/70