Use the `system.system` id from the config
Even though we expect there to only be one of these, we can still do the
more correct thing and use the actual id that was configured.

This is also more in line with what the other data sources do.

Branch: joneshf/use-the-system-system-id-from-the-config
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/73