Make a helper for the id attribute
We seem to be doing the same thing for the id attribute everywhere.
Instead of copy/pasting this more (and to keep the ids consistent) we
pull this helper to a common location and use it in all the locations.

Branch: joneshf/make-a-helper-for-the-id-attribute
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/75