Abstract the behavior of Data Sources
Since all of the Data Sources are doing the same stuff, we pull out the
behavior to a common location that we can use in multiple places. We
then initialize it with the pieces that are different: a getter for the
id, the schema stuff, and the UCI stuff.

Then, we can simplify the actual Data Sources to almost nothing. We
still keep them around because testing their behavior is different for
each one. But this might be as minimal as we can get things.

Branch: joneshf/abstract-the-behavior-of-data-sources
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/76