Abstract the behavior of Resources
Much like the change for Data Sources, the Resources we have are all
doing the same stuff. We follow a similar pattern for Resources and put
that behavior in a common location, and use it in multiple places.

This also whittled the actual Resource files down to almost nothing.

Branch: joneshf/abstract-the-behavior-of-resources
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/77