Add import docs for each resource
It can be non-trivial to figure out how to import each of these
resources. We provide a little bit of help to get people one their way.

It'd be better to have a more robust way to figure this stuff out, but
short of writing a program that does this work ourselves, it's not clear
that there is a better way. That might not be a bad idea, to be honest.