Relax provider's dependency on environment
We do want to support reading default values from the environment, but
we don't need to have a hard dependency on using the actual environment
in the provider's implementation. Instead, we can describe what
dependency we actually have (looking up a value in the environment) and
let the callers pass their version of the dependency in.

This really doesn't do much at the moment, but it paves the way for
writing a test that can still run in parallel even though it's doing
things to the "environment." We'll add that test in the future.

Branch: joneshf/relax-provider-s-dependency-on-environment
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/81