Fix port lookup in the environment
We had the boolean flipped when deciding whether or not to use the
parsed value. We flip it the other way, so it should work properly now.

This is another example of how having sum types would have prevented the
bug from even happening in the first place. If we could represent the
result of `strconv.Atoi` as a sum of `int` or `error` (rather than a
product of both `int` and `error`), we wouldn't have needed to check
whether an `error` was `nil` or not. In fact, if sum types permeated
through the language more, we likely wouldn't even have `nil` as a
possible value of `error` to check in the first place. In any case, sums
could've prevented another bug.

Branch: joneshf/add-a-test-that-should-pass-using-the-environment
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/82