Switch over to using `nix` flakes
It seems flakes are going to take over the world, so we're going to try
them out for this.

The thing that seems actually interesting is that since flakes require
the `nix` command (rather than the set of `nix-*` commands we were
using), it means we might be able to use the `--store` argument to
specify a directory that can be used for caching. The idea is that maybe
we can get around the issues pointed out in
https://github.com/actions/cache/issues/749 and
https://github.com/actions/cache/issues/324#issuecomment-873296255.

That's the hope, anyway. Let's see if we can make it happen.