Attempt to cache nix dependencies
We specify a local directory for the nix store `~/nix`, and use it in
the following `nix` commands. We toss up a simple `echo` command to
prime the dependencies that we've got, so we make sure to cache stuff
even if one of the further steps fails.

This might actually be a misstep, and we could just go with the implicit
behavior of the other steps pulling down dependencies for us. We'll see
how it works out.