Remove priming of dependencies
The other commands should do all we need for getting the correct
dependencies downloaded and into the directory for caching. It should be
enough to rely on those instead of having an explicit command that
forces the downloads.

This commit also serves as a check that the cache is actually working.
Something was uploaded here:
https://github.com/joneshf/terraform-provider-openwrt/actions/caches
(for posterity the key is
`Linux-5001d141c298c6350d540e9f8eeaa7aacbd54b52c9dc9bc395db14311eee6957`
and it says that 900 MB were cached), so we should have a cache hit.