Switch over to using `nix` flakes
It seems flakes are going to take over the world, so we're going to try
them out for this.

The thing that is actually interesting about this change is that since
flakes require the `nix` command (rather than the set of `nix-*`
commands we were using), we can use the `--store` argument to specify a
directory that can be used for caching. This way, we can get around the
issues pointed out in https://github.com/actions/cache/issues/749 and
https://github.com/actions/cache/issues/324#issuecomment-873296255.

We'll move on to caching other stuff after this. But for now, just
caching the `nix` stuff seems to knock about a minute off of CI. It was
between 3m 30s and 4m. Now it's sitting around 2m 45s. That's ~20%
decrease in CI time! Here's hoping future caching efforts yield other
useful gains.

Branch: joneshf/switch-over-to-using-nix-flakes
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/83