Use a more specific cache for nix stuff
Before we add a cache for other things, we should be a little more
specific with the nix cache. We toss the word `nix` into the key, so
we're less likely to conflict with other stuff.

We also make the restore keys fallback to the closest cache they can
find. We don't want to have to redo everything if a cache miss happens
because we added or removed a single dependency.

Branch: joneshf/use-a-more-specific-cache-for-nix-stuff
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/84