Cache `go` stuff
We follow what we're doing for `nix` and cache the dependencies and
build artifacts for `go`.

We end up having to use `go.mod` as a hash as well because it is the
only place we have the go version written down anywhere, and we want
that to be part of the cache invalidation strategy.