Cache `go` stuff
We follow what we're doing for `nix` and cache the dependencies and
build artifacts for `go`.

We end up having to use `go.mod` as a hash as well because it is the
only place we have the go version written down anywhere, and we want
that to be part of the cache invalidation strategy.

This change shaves a large chunk off CI time. For posterity, the first
commit here was 3m 9s. The second commit here was 1m 29s. That's more
than 50% of the CI!

Branch: joneshf/cache-go-stuff
Pull-Request: https://github.com/joneshf/terraform-provider-openwrt/pull/85